// src/app/model/topic.ts
export enum Topic {
  CONNECTION = 'connection',
  DISCONNECT = 'disconnect',
  FINDLOBBY = 'findlobby',
  LISTLOBBIES = 'listlobbies',
  NEWLOBBY = 'newlobby',
  JOINLOBBY = 'joinlobby',
  LOBBYINFO = 'lobbyinfo',
  STARTGAME = 'startgame',
  JOINGAME = 'joingame',
  GAMEDATA = 'gamedata',
  BROADCAST = 'broadcast',
  HOME = 'home',
  LOBBYSETTINGS = 'lobbysettings',
  MAFIACHAT = 'mafiachat',
  REFRESHDATA = 'refreshdata',

}